tenderize: v2 adapater, rename v1 folder
